requirements.txt
# Azure ML
azureml-core==1.28.0
azureml-train-core==1.28.0
azureml-dataprep

# basic
pandas
matplotlib
jupyter

mecab-python3==1.0.3
unidic-lite==1.0.8
gensim==4.0.1
python-Levenshtein==0.12.2

torch==1.8.1
pytorch-lightning==1.3.1
mlflow